Update to Mailer.php to fix attach function
I discovered an issue where the mailer was not attaching files provided, putting up this PR as this is what resolved it for me. 

Details:
`$data` returns an associative array which `list()` doesn't know how to handle, which causes it to not actually attach a file provided. 

Adding array_values around $data on line 276 will return $data as an indexed array and seems to resolve the issue and allows files to be attached.